# server/docs/swagger.yaml
definitions:
  main.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  openai.AudioResponse:
    properties:
      text:
        description: Text is the text used to generate the audio.
        type: string
    type: object
...
  openai.DeleteFileResponse:
    properties:
      deleted:
        description: Deleted is a boolean indicating whether the file was successfully
          deleted.
        type: boolean
      id:
        description: ID is the ID of the deleted file.
        type: string
      object:
        description: Object is the type of object for the response.
        type: string
    type: object
  openai.DialogRequest:
    properties:
      input:
        description: Input is the input for the dialog.
        type: string
      instruction:
        description: Instruction is the instruction for the dialog.
        type: string
    type: object
...
  openai.EmbeddingRequest:
    properties:
      input:
        description: Input is the input text to generate the embedding for.
        type: string
      model:
        description: Model is the ID of the model to use for generating the embedding.
        type: string
    type: object
...
  openai.FileInfo:
    properties:
      bytes:
        description: Bytes 是文件的大小（以字节为单位）。
        type: integer
      created_at:
        description: CreatedAt 是文件创建的时间戳。
        type: integer
      filename:
        description: Filename 是文件的名称。
        type: string
      id:
        description: ID 是文件的唯一标识符。
        type: string
      object:
        description: Object 是响应的对象类型。
        type: string
      purpose:
        description: Purpose 是文件的用途。
        type: string
    type: object
...
info:
  contact: {}
paths:
  /openai/api/v1/audio/transcriptions:
    post:
      consumes:
      - multipart/form-data
      description: Transcribe an audio file to text
      parameters:
      - description: Audio file to transcribe
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.AudioResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Transcribe audio file
  /openai/api/v1/audio/translations:
    post:
      consumes:
      - multipart/form-data
      description: Translate an audio file to text
      parameters:
      - description: Audio file to translate
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.AudioResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Translate audio file
  /openai/api/v1/chat:
    post:
      consumes:
      - application/json
      description: 使用OpenAI的API完成聊天提示
      parameters:
      - description: 聊天提示的输入
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/openai.DialogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.ChatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
  /openai/api/v1/chat/edit:
    post:
      consumes:
      - application/json
      description: Edit a chat prompt using OpenAI's API
      parameters:
      - description: Input for chat prompt
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/openai.DialogRequest'
      - description: Instruction for chat prompt editing
        in: body
        name: instruction
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.EditChatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Edit a chat prompt
  /openai/api/v1/completions:
    post:
      consumes:
      - application/json
      description: Complete a text prompt using OpenAI's API
      parameters:
      - description: Input for text prompt
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/openai.DialogRequest'
      - description: Name of the model to use for completion
        in: query
        name: model
        required: true
        type: string
      - description: Sampling temperature to use for completion
        in: query
        name: temperature
        required: true
        type: number
      - description: Maximum number of tokens to generate for completion
        in: query
        name: max_tokens
        required: true
        type: integer
      - description: Number of completions to generate
        in: query
        name: "n"
        required: true
        type: integer
      - description: Sequence to stop generation at
        in: query
        name: stop
        required: true
        type: string
      - description: Sequence to force into the generated text
        in: query
        name: presence
        required: true
        type: string
      - description: Frequency penalty to use for completion
        in: query
        name: frequency_penalty
        required: true
        type: number
      - description: Presence penalty to use for completion
        in: query
        name: presence_penalty
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.CompletionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Complete a text prompt
  /openai/api/v1/embeddings:
    post:
      consumes:
      - application/json
      description: Get embeddings for a given input
      parameters:
      - description: Input for which embeddings are to be generated
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/openai.EmbeddingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.EmbeddingResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Get embeddings
  /openai/api/v1/files/{file_id}:
    delete:
      consumes:
      - application/json
      description: Delete a fine-tuned file
      parameters:
      - description: File ID
        in: path
        name: file_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.DeleteFileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Delete a file
    get:
      consumes:
      - application/json
      description: Get information about a fine-tuned file
      parameters:
      - description: File ID
        in: path
        name: file_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.FileInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Get file info
  /openai/api/v1/files/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload a file to be fine-tuned
      parameters:
      - description: File to be uploaded
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.FileInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Upload a file
  /openai/api/v1/fine-tunes:
    get:
      consumes:
      - application/json
      description: Get a list of all fine-tune jobs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.FineTuneJobList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Get fine-tune job list
  /openai/api/v1/fine-tunes/{file_id}:
    post:
      consumes:
      - application/json
      description: Create a fine-tune job using a file ID
      parameters:
      - description: File ID
        in: path
        name: file_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.FineTuneJob'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Create a fine-tune job
  /openai/api/v1/fine-tunes/{fine_tune_id}:
    delete:
      consumes:
      - application/json
      description: Delete a fine-tune job using a fine-tune job ID
      parameters:
      - description: Fine-tune job ID
        in: path
        name: fine_tune_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.JobDeleteInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Delete a fine-tune job
    get:
      consumes:
      - application/json
      description: Get information about a fine-tune job
      parameters:
      - description: Fine-tune job ID
        in: path
        name: fine_tune_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.FineTuneJob'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Get fine-tune job
  /openai/api/v1/fine-tunes/{fine_tune_id}/cancel:
    post:
      consumes:
      - application/json
      description: Cancel a fine-tune job using a fine-tune job ID
      parameters:
      - description: Fine-tune job ID
        in: path
        name: fine_tune_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.FineTuneJob'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Cancel a fine-tune job
  /openai/api/v1/fine-tunes/{fine_tune_id}/events:
    get:
      consumes:
      - application/json
      description: Get events for a fine-tune job
      parameters:
      - description: Fine-tune job ID
        in: path
        name: fine_tune_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.FineTuneJobEventList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Get fine-tune job events
  /openai/api/v1/images:
    get:
      consumes:
      - application/json
      description: Generate an image using OpenAI's DALL-E API
      parameters:
      - description: Model to use for image generation
        in: query
        name: model
        required: true
        type: string
      - description: Number of images to generate
        in: query
        name: "n"
        required: true
        type: integer
      - description: Size of the image to generate
        in: query
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.ImageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Generate an image
      tags:
      - Images
  /openai/api/v1/images/edit:
    post:
      consumes:
      - multipart/form-data
      description: Edit an image using OpenAI's DALL-E API
      parameters:
      - description: Image to edit
        in: formData
        name: image
        required: true
        type: file
      - description: Prompt for image editing
        in: formData
        name: prompt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/openai.ImageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Edit an image using OpenAI's DALL-E API
      tags:
      - Images
...
swagger: "2.0"